Put in a fix to a problem pointed out by "Ing. Jozef Sakalos" <jsakalos@ba.success.sk> It's a signed/unsigned comparison problem.
CVS patchset: 2126
CVS date: 2003/07/21 13:12:30

+++ lib/clplumbing/ipcsocket.c
@@ -798,7 +798,7 @@ socket_resume_io_write(struct IPC_CHANNEL *ch, gboolean* started)
 		} while(sendrc < 0
 		&&	(errno == EAGAIN ? (cl_shortsleep(), TRUE) : FALSE));
 
-		if (sendrc != msg->msg_len) {
+		if (sendrc != (int)msg->msg_len) {
 			cl_perror("Sent %d byte message body: rc = %d"
 			,	(int)msg->msg_len, sendrc);
 		}
